requirements.txt
Pillow
opencv-python
dlib           
numpy
PySide6